% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{United States}
  \item{iso2, iso3}{US, USA}
  \item{year}{}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
fire
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
