% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{dataset to sample from}

\item{fun}{function to run on the samples}

\item{alpha}{1-alpha CI}

\item{...}{extra arguments}
}
\value{
a list containing the CI, function, dataset, and xstat
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(x=1:10)}
}
