% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of random values per sample}

\item{iter}{number of samples}

\item{lambda}{lambda to use in poisson}

\item{...}{extra arguments}
}
\value{
a plot
}
\description{
mycltp
}
\examples{
\dontrun{mycltp(100, 10000)}
}
