% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to apply across x}

\item{x}{dataset}

\item{param}{parameters to feed into lfun}

\item{...}{extra arguments}
}
\value{
list of answers to the maximum likelihood
}
\description{
mymaxlik
}
