% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the distribution}

\item{sigma}{the standard deviation}

\item{a}{maximum of interval (-Inf, a)}
}
\value{
a normal distribution with (-Inf, a) highlighted
}
\description{
Title
}
\examples{
\dontrun{myncurve(5, 10, 3)}
}
