% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{number of times to sample}

\item{iter}{number of iterations/graphs to generate}

\item{time}{amount of time to wait between each generation}
}
\description{
mysample
}
\examples{
\dontrun{mysample(n=1000,iter=30,time=1)}
}
