% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the plane}

\item{gamma}{probability of an overbook, i.e. too many people show up}

\item{p}{probability a passenger shows up}
}
\value{
a list containing the discrete and continuous answers, as well as the parameters passed in
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N=400,gamma = 0.02, p = 0.95)}
}
