% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadf.R
\name{quadf}
\alias{quadf}
\title{Quadratic Function}
\usage{
quadf(x, coef)
}
\arguments{
\item{x}{X variable}

\item{coef}{Array of 3 coefficients for a quadratic in the order [c, b, a]}
}
\value{
The function evaluated at X.
}
\description{
Quadratic Function
}
\examples{
\dontrun{myf(15, coef = c(2, 5, 7))}
}
